+++ src/audio_metadata/formats/vorbis_comments.py
@@ -11,7 +11,10 @@ from collections import defaultdict
 from attr import attrib, attrs
 from tbm_utils import datareader
 
-from ..exceptions import FormatError
+from ..exceptions import (
+	FormatError,
+	TagError,
+)
 from ..models import (
 	Tag,
 	Tags,
@@ -25,6 +28,17 @@ from ..models import (
 class VorbisComment(Tag):
 	name = attrib(converter=lambda n: n.lower())
 
+	@staticmethod
+	def _validate_name(name):
+		return all(
+			(
+				char >= ' '
+				and char <= '}'
+				and char != '='
+			)
+			for char in name
+		)
+
 	@datareader
 	@classmethod
 	def parse(cls, data):
@@ -36,6 +50,9 @@ class VorbisComment(Tag):
 
 		name, value = comment.split('=', 1)
 
+		if not cls._validate_name(name):
+			raise TagError(f"Invalid character in Vorbis comment name: ``{name}``.")
+
 		return cls(
 			name=name,
 			value=value,
@@ -44,6 +61,13 @@ class VorbisComment(Tag):
 
 # TODO: Number frames.
 class VorbisComments(Tags):
+	def __init__(self, *args, **kwargs):
+		super().__init__(*args, **kwargs)
+
+		for key in self.keys():
+			if not VorbisComment._validate_name(key):
+				raise TagError(f"Invalid character in Vorbis comment name: ``{key}``.")
+
 	@datareader
 	@classmethod
 	def parse(cls, data):